Add a warning about duplicate USB pullup for STM32F4 boards
This is suboptimal hardware design, which might need some special care
on the software side. This noticed in the Arduino_Core_STM32, see
https://github.com/stm32duino/Arduino_Core_STM32/issues/1029

For these boards, a schematic is available to confirm the resistor:
 - STM32F407VET6-STM32-F4VE-V2.0
 - STM32F407ZGT6-VCC-GND-Large
 - STM32F407VET6-VCC-GND-Small

These boards have not schematic, but there is a 1.5kΩ resistor clearly
visible in the USB data path that is almost certainly a fixed pullup:
 - STM32F407VET6-Euse-M4-DEMO-Medium
 - STM32F407VGT6-SR-Board

These boards have a pullup, but it is switched by a transistor. Since
there are also internal pullups, a warning is still in order, but using
slightly different wording:
 - STM32F401RCT6-STM32F-Core-Board
 - STM32F407ZGT6-STM32F-Core-Board

All other F4 boards have no such resistor in the schematic or visible in
the images. Other series might also have this problem, but were not
checked.